+++ lib/noop_backup/notifiers/stdout.rb
@@ -1,7 +1,7 @@
 module NoopBackup::Notifiers
   class Stdout
-    def notify(text)
-      puts text
+    def notify(result)
+      result.messages.each { |message| puts message }
     end
   end
 end