+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Make s3 upload stream part size and thread count configurable
+- Sentinel integration
 
 ## [0.2.0] - 2026-07-04